Split aiplatform.constants into multiple modules
