+++ tests/unit/aiplatform/test_datasets.py
@@ -30,6 +30,7 @@ from google.auth import credentials as auth_credentials
 from google.cloud import aiplatform
 from google.cloud.aiplatform import base
 from google.cloud.aiplatform import compat
+from google.cloud.aiplatform.constants import base as constants
 from google.cloud.aiplatform import datasets
 from google.cloud.aiplatform import initializer
 from google.cloud.aiplatform import schema
@@ -514,7 +515,7 @@ class TestDataset:
 
         assert (
             ds.api_client._clients[compat.DEFAULT_VERSION]._client_options.api_endpoint
-            == f"{_TEST_LOCATION}-{aiplatform.constants.API_BASE_PATH}"
+            == f"{_TEST_LOCATION}-{constants.API_BASE_PATH}"
         )
 
         assert _TEST_ALT_LOCATION != _TEST_LOCATION
